orchestrator/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
docker>=6.1.0
pyyaml>=6.0.1
jsonschema>=4.19.0
pydantic>=2.4.0
